requirements.txt
groq>=0.9.0
sentence-transformers>=2.2.0
bertopic>=0.16.0
chromadb>=0.5.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
langdetect>=1.0.9
langid>=1.1.6
spacy>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
hdbscan>=0.8.29
umap-learn>=0.5.0
jinja2>=3.1.0
python-dotenv>=1.0.0
httpx>=0.24.0
click>=8.1.0
pytz>=2024.1
pyyaml>=6.0
flask>=3.0.0
pytest>=7.4.0
